Classic Snake Game
The Snake Game project is a classic arcade-style game developed in Java using the Swing library. The objective of the game is to control a snake, guiding it to consume randomly generated food (enemy) while avoiding collisions with the game borders and the snake's own body. The project provides a simple and entertaining gaming experience, showcasing key concepts in Java programming and graphical user interface (GUI) development.